web/web-lobby/src/main/java/com/yazino/web/payment/googlecheckout/v3/GooglePurchase.java
package com.yazino.web.payment.googlecheckout.v3;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.databind.annotation.JsonSerialize;
import com.yazino.web.payment.Purchase;
import org.apache.commons.lang3.builder.EqualsBuilder;
import org.apache.commons.lang3.builder.HashCodeBuilder;
import org.apache.commons.lang3.builder.ToStringBuilder;
import com.fasterxml.jackson.annotation.JsonProperty;

@JsonIgnoreProperties(ignoreUnknown = true)
@JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
public class GooglePurchase extends Purchase {

    @JsonProperty
    private Boolean canConsume;

    public boolean canConsume() {
        return canConsume;
    }

    public void setCanConsume(Boolean canConsume) {
        this.canConsume = canConsume;
    }

    @Override
    public boolean equals(Object obj) {
        if (obj == null) {
            return false;
        }
        if (obj == this) {
            return true;
        }
        if (obj.getClass() != getClass()) {
            return false;
        }
        GooglePurchase rhs = (GooglePurchase) obj;
        return new EqualsBuilder()
                .appendSuper(super.equals(obj))
                .append(this.canConsume, rhs.canConsume)
                .isEquals();
    }

    @Override
    public int hashCode() {
        return new HashCodeBuilder()
                .appendSuper(super.hashCode())
                .append(canConsume)
                .toHashCode();
    }

    @Override
    public String toString() {
        return new ToStringBuilder(this)
                .appendSuper(super.toString())
                .append("canConsume", canConsume)
                .toString();
    }
}
